pkg/util/trivy: move containerd trivy code to separate file
